backend/requirements.txt
Flask>=2.2.0
flask-cors>=3.0.10
pandas>=2.0
sqlalchemy>=1.4
requests>=2.0
pymysql>=1.0.2
psycopg2-binary>=2.9.0
PyJWT>=2.4.0